+++ src/main/java/com/univ/kanban/dto/LoginDto.java
@@ -1,0 +1,21 @@
+package com.univ.kanban.dto;
+
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class LoginDto {
+
+    @NotEmpty
+    @NotNull
+    private String email;
+
+    @NotEmpty
+    @NotNull
+    private String password;
+}
