+++ src/main/java/com/univ/kanban/dto/UserDto.java
@@ -1,0 +1,29 @@
+package com.univ.kanban.dto;
+
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class UserDto {
+
+    @NotNull
+    @NotEmpty
+    private String prenom;
+
+    @NotNull
+    @NotEmpty
+    private String nom;
+
+    @NotNull
+    @NotEmpty
+    private String email;
+
+    @NotNull
+    @NotEmpty
+    private String password;
+}
