+++ src/main/java/com/univ/kanban/repositories/ColumnRepository.java
@@ -1,0 +1,7 @@
+package com.univ.kanban.repositories;
+
+import com.univ.kanban.models.KanbanColumn;
+import org.springframework.data.repository.CrudRepository;
+
+public interface ColumnRepository extends CrudRepository<KanbanColumn, Long>{
+}
