+++ src/main/java/com/univ/kanban/services/KanbanService.java
@@ -1,46 +1,46 @@
-package com.univ.kandan.service;
+package com.univ.kanban.services;
 
 import java.util.Set;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
-import com.univ.kandan.dto.ColumnDto;
-import com.univ.kandan.dto.KanbanDto;
-import com.univ.kandan.model.Column;
-import com.univ.kandan.model.Kanban;
-import com.univ.kandan.repository.ColumnRepository;
-import com.univ.kandan.repository.KanbanRepository;
+import com.univ.kanban.dto.ColumnDto;
+import com.univ.kanban.dto.KanbanDto;
+import com.univ.kanban.models.KanbanColumn;
+import com.univ.kanban.models.Kanban;
+import com.univ.kanban.repositories.ColumnRepository;
+import com.univ.kanban.repositories.KanbanRepository;
 
 @Service
 public class KanbanService {
 
     private final ModelMapper modelMapper;
-    private final KanbanRepository kanbanRepostitory;
+    private final KanbanRepository kanbanRepository;
     private final ColumnRepository columnRepository;
 
     public KanbanService(KanbanRepository kanbanRepository, ColumnRepository columnRepository, ModelMapper modelMapper) {
-        this.kanbanRepostitory = kanbanRepository;
+        this.kanbanRepository = kanbanRepository;
         this.columnRepository = columnRepository;
         this.modelMapper = modelMapper;
     }
 
     public Set<Kanban> findAllIsPublic() {
-        return kanbanRepostitory.findAllByIsPublic(true);
+        return kanbanRepository.findAllByIsPublic(true);
     }
 
     public Set<Kanban> findAllByUser(Long userId) {
-        return kanbanRepostitory.findAllByUser(userId);
+        return kanbanRepository.findAllByUser(userId);
     }
 
     public Kanban findById(Long id) {
-        return kanbanRepostitory.findById(id);
+        return kanbanRepository.findById(id);
     }
 
     public Kanban create(KanbanDto kanbanDto) {
         Kanban kanban = new Kanban();
         beforeCreate(kanbanDto, kanban);
-        kanban = kanbanRepostitory.save(kanban);
+        kanban = kanbanRepository.save(kanban);
         afterCreate(kanbanDto, kanban);
         return kanban;
     }
@@ -51,7 +51,7 @@ public class KanbanService {
 
     public void afterCreate(KanbanDto kanbanDto, Kanban kanban) {
         for (ColumnDto col : kanbanDto.getColumns()) {
-            Column column = modelMapper.map(col, Column.class);
+            KanbanColumn column = modelMapper.map(col, KanbanColumn.class);
             column.setKanban(kanban);
             columnRepository.save(column);
         }
