--- src/main/java/com/univ/kandan/Exceptions/InvalidCredentialsException.java
@@ -1,15 +1,0 @@
-package com.univ.kandan.exceptions;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class InvalidCredentialsException extends AuthenticationException {
-
-  public InvalidCredentialsException(String msg) {
-    super(msg);
-  }
-
-  public InvalidCredentialsException() {
-    super(null);
-  }
-
-}
