--- src/main/java/com/univ/kandan/Repository/ColumnRepository.java
@@ -1,9 +1,0 @@
-package com.univ.kandan.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.univ.kandan.model.Column;
-
-public interface ColumnRepository extends CrudRepository<Column, Long>{
-    
-}
