--- src/main/java/com/univ/kandan/dto/LoginDto.java
@@ -1,31 +1,0 @@
-package com.univ.kandan.dto;
-
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-
-public class LoginDto {
-
-  @NotEmpty
-  @NotNull
-  private String email;
-
-  @NotEmpty
-  @NotNull
-  private String password;
-
-  public String getEmail() {
-    return email;
-  }
-
-  public String getPassword() {
-    return password;
-  }
-
-  public void setEmail(String email) {
-    this.email = email;
-  }
-
-  public void setPassword(String password) {
-    this.password = password;
-  }
-}
