+++ src/main/resources/application.properties
@@ -1,11 +1,7 @@
-server.port = 8080
-spring.datasource.url=jdbc:mysql://localhost:3306/kanban
-spring.datasource.username=root
-spring.datasource.password=
+server.port=8080
+spring.datasource.url=jdbc:mysql://localhost:3306/eikjos
+spring.datasource.username=eikjos
+spring.datasource.password=eikjos
 spring.application.name=kanban
 spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.database-platform=org.hibernate.dialect.MySQL5Dialect
-spring.jpa.open-in-view =false
-spring.thymeleaf.enabled=true
-spring.main.allow-bean-definition-overriding=true
+spring.jpa.properties.hibernate.globally_quoted_identifiers=true
