+++ src/main/resources/templates/kanban/kanban.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="fr" xmlns:th="https://www.thymeleaf.org">
+<head>
+    <title>Kandan</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <script src="https://cdn.tailwindcss.com"></script>
+</head>
+<body>
+<header class="bg-emerald-700 py-5 flex justify-center items-center">
+    <h1 class="text-5xl font-bold text-white ml-3">KANBAN</h1>
+    <div sec:authorize="isAuthenticated()" class="absolute right-10 bg-white p-3 rounded-xl border-2 border-black">
+        <form th:action="@{/logout}" method="post">
+            <button type="submit">Déconnexion</button>
+        </form>
+    </div>
+</header>
+<div class="flex flex-row justify-center items-baseline">
+    <div class="w-1/3 justify-center mt-10 text-2xl bg-slate-500 text-center text-white p-2 rounded-lg">
+        <h1 th:text="${kanban.nom}"></h1>
+    </div>
+</div>
+</body>
+</html>
